cherrypick: move tests
